A patch from Leon "incorrect url of root webapp in menu" https://issues.apache.org/jira/browse/OFBIZ-4368
If I enable the "app-bar-display" for root webapp which is mounted to "/", the url of root webapp in menu is "https://control/main" and not the correct "https://localhost:8443/".

That's because, e.g. in themes/tomahawk/includes/appbarOpen.ftl, it uses expression  <#if thisApp != "/"> to judge whether the app is root or not, while the variable thisApp is java String and it'll be encoded as iso code "/", so the "if" expression mentioned above will always return false.

In the patch it use StringUtil.wrapString to wrap variable theApp to avoid encode.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/themes/flatgrey@1158124 13f79535-47bb-0310-9956-ffa450edef68