An improved patch from Pierre Gaudin for "JGrowl to display alerts " https://issues.apache.org/jira/browse/OFBIZ-6023
I've been working on Jgrowl integration for a project. (https://github.com/stanlemon/jGrowl)
This allowS to display system's messages in a popup at the top of the windows even if you scroll. This is also useful for ajax call because there is no need to scroll to the top of the screen to view message.
Message can stick to the screen until it is intentionally closed by the user or not.

jleroux? as already explained in the Jira: to have it working in trunk HEAD I had to make some changes

    I put the jgrowl loading in GlobalActions rather than GlobalDecorator
    in showjGrowl() js function
        I Added a test for if (contentMessages.length) because of an error in lookups
        I did some formatting in function showjGrowl()
        I used jQuery instead of $ (same for expandAllP() and expandAll() functions)


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/themes/flatgrey@1709679 13f79535-47bb-0310-9956-ffa450edef68