Improves: Use ignore_missing option of the <#include Freemarker directive when fixed (OFBIZ-8292)
This is a no functional change except for the flatgrey and blueligth themes where svn and git info are added

Working on OFBIZ-8250 and after Deepak at OFBIZ-7942 (too bad I missed that :/)  I found that the ignore_missing option of the <#include Freemarker directive does not work. I reported to the Freemarker incubating project.
Hopefully this will be fixed. It will then remove the need of creating empty files for the svn and git info footers when building. 

In the meantime this provides the change in the themes footers.
I have also added the svn and git info to the flatgrey footer. The result is barely legible there but is also used (include) by the bluelight theme where it's OK
I have also formatted the related div where it was barely legible (too large)


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk/themes/flatgrey@1761392 13f79535-47bb-0310-9956-ffa450edef68